requirements.txt
streamlit
python-dotenv
pandas

pdfplumber
python-docx

langchain
langchain-community
langchain-core

langchain-text-splitters
chromadb
sentence-transformers
torch

groq